editoast: move Endpoint and TrackEndpoint to editoast_schemas
