+++ editoast/editoast_schemas/src/infra.rs
@@ -1,8 +1,10 @@
 mod applicable_directions;
 mod direction;
 mod directional_track_range;
+mod endpoint;
 mod side;
 mod sign;
+mod track_endpoint;
 mod track_location;
 mod track_offset;
 mod track_range;
@@ -11,8 +13,10 @@ mod waypoint;
 pub use applicable_directions::ApplicableDirections;
 pub use direction::Direction;
 pub use directional_track_range::DirectionalTrackRange;
+pub use endpoint::Endpoint;
 pub use side::Side;
 pub use sign::Sign;
+pub use track_endpoint::TrackEndpoint;
 pub use track_location::TrackLocation;
 pub use track_offset::TrackOffset;
 pub use track_range::TrackRange;
