+++ editoast/editoast_schemas/src/infra/endpoint.rs
@@ -1,0 +1,9 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Hash)]
+#[serde(rename_all = "UPPERCASE")]
+pub enum Endpoint {
+    Begin,
+    End,
+}
