+++ editoast/src/converters/generate_routes.rs
@@ -9,6 +9,8 @@ use std::collections::HashMap;
 use crate::schema::*;
 use editoast_common::Identifier;
 use editoast_schemas::infra::Direction;
+use editoast_schemas::infra::Endpoint;
+use editoast_schemas::infra::TrackEndpoint;
 use editoast_schemas::infra::Waypoint;
 use editoast_schemas::primitives::OSRDIdentified;
