+++ editoast/src/converters/osm_to_railjson.rs
@@ -127,13 +127,15 @@ pub fn parse_osm(osm_pbf_in: PathBuf) -> Result<RailJson, Box<dyn Error + Send +
 
 #[cfg(test)]
 mod tests {
+    use editoast_common::Identifier;
     use editoast_schemas::infra::ApplicableDirections;
+    use editoast_schemas::infra::TrackEndpoint;
     use std::collections::HashMap;
 
     use super::parse_osm;
     use crate::converters::*;
     use crate::schema::*;
-    use editoast_common::Identifier;
+
     #[test]
     fn convert_osm_to_railjson() {
         let output = tempfile::NamedTempFile::new().unwrap();
