+++ editoast/src/schema/route.rs
@@ -4,10 +4,8 @@ use derivative::Derivative;
 use serde::Deserialize;
 use serde::Serialize;
 
-use super::Endpoint;
 use super::OSRDIdentified;
 use super::ObjectType;
-use super::TrackEndpoint;
 use crate::infra_cache::Cache;
 use crate::infra_cache::Graph;
 use crate::infra_cache::InfraCache;
@@ -15,6 +13,8 @@ use crate::infra_cache::ObjectCache;
 use editoast_common::Identifier;
 use editoast_schemas::infra::Direction;
 use editoast_schemas::infra::DirectionalTrackRange;
+use editoast_schemas::infra::Endpoint;
+use editoast_schemas::infra::TrackEndpoint;
 use editoast_schemas::infra::Waypoint;
 use editoast_schemas::primitives::OSRDTyped;
