+++ editoast/src/schema/switch.rs
@@ -6,11 +6,11 @@ use serde::Serialize;
 
 use super::OSRDIdentified;
 use super::ObjectType;
-use super::TrackEndpoint;
 use crate::infra_cache::Cache;
 use crate::infra_cache::ObjectCache;
 use editoast_common::Identifier;
 use editoast_common::NonBlankString;
+use editoast_schemas::infra::TrackEndpoint;
 use editoast_schemas::primitives::OSRDTyped;
 
 #[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq)]
