+++ editoast/src/schema/track_section.rs
@@ -5,15 +5,15 @@ use geos::geojson::Value::LineString;
 use serde::Deserialize;
 use serde::Serialize;
 
-use super::Endpoint;
 use super::OSRDIdentified;
 use super::ObjectType;
-use super::TrackEndpoint;
 use crate::infra_cache::Cache;
 use crate::infra_cache::ObjectCache;
 use crate::map::BoundingBox;
 use editoast_common::Identifier;
 use editoast_common::NonBlankString;
+use editoast_schemas::infra::Endpoint;
+use editoast_schemas::infra::TrackEndpoint;
 use editoast_schemas::primitives::OSRDTyped;
 
 #[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq)]
