+++ editoast/src/views/infra/auto_fixes/track_section.rs
@@ -7,11 +7,11 @@ use super::new_ref_fix_create_pair;
 use super::Fix;
 use crate::infra_cache::operation::RailjsonObject;
 use crate::schema::BufferStop;
-use crate::schema::Endpoint;
 use crate::schema::InfraError;
 use crate::schema::InfraErrorType;
 use crate::schema::TrackSectionCache;
 use editoast_common::Identifier;
+use editoast_schemas::infra::Endpoint;
 use editoast_schemas::primitives::OSRDIdentified as _;
 use editoast_schemas::primitives::OSRDObject as _;
 use editoast_schemas::primitives::ObjectRef;
